M    hrl_camera/camera.py M    hrl_camera/hrl_camera.py M    hrl_camera/camera_setup_lib.py M    hrl_camera/hrl_ros_camera.py
debugging why starting a second camera process kills the previous one.
It turns out that the opencv CaptureFromCAM is at fault.
I am using the param server to put the ids of all the cameras to
prevent the camera_setup_lib from accessing the entire dc1394 bus.

This commit is purely for documentation. Need to start all the cameras
connected to a computer through one process.


git-svn-id: https://gt-ros-pkg.googlecode.com/svn/trunk@358 5fc10e88-4e10-11de-b850-9977aa68d080