the camera1394 package (part of the official ros stacks) has deprecated most of this package.
what is useful is image_listener.py which is an example of how to get
images published by some node of the image pipeline (ROS IplImage
message) to opencv format in python, using the opencv bridge.

look at robot1/launch/cody/utm_camera.launch for a launch file that
uses the image pipeline with a firefly camera.


git-svn-id: https://gt-ros-pkg.googlecode.com/svn/trunk@668 5fc10e88-4e10-11de-b850-9977aa68d080